+++ src/rules/at-function-pattern/index.js
@@ -24,7 +24,7 @@ export default function (pattern) {
       if (decl.name !== "function") { return }
 
       // Stripping the function of its arguments
-      const funcName = decl.params.replace(/(\s*?)\(.*?\)/g, "")
+      const funcName = decl.params.replace(/(\s*?)\((?:\s|\S)*?\)/g, "")
       if (regexpPattern.test(funcName)) { return }
 
       utils.report({